# src/users/entities/user.entity.ts
export class User {
    id: string;
    email: string;
    firstName: string;
    lastName: string;
    phoneNumber: string;
    dateOfBirth: Date;
    identityNumber?: string;
    identityType?: string;
    createdAt: Date;
    updatedAt: Date;
}